requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-bower==5.0.4
django-filter==0.9.2
django-toolbelt==0.0.1
djangorestframework==3.1.1
gunicorn==19.4.5
Markdown==2.6.2
Pillow==3.1.0
psycopg2==2.4.2
six==1.10.0
static3==0.6.1
wheel==0.24.0
